feat: refactor caching for craft add-intro
